# fast-lane-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Fast Lane Delivery API",
        "description": "API Documentation for Fast Lane Delivery and Guesthouse Booking service",
        "contact": {
            "email": "admin@fastlanedelivery.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/reviews": {
            "get": {
                "tags": [
                    "Reviews"
                ],
                "summary": "List reviews",
                "description": "Retrieves a list of reviews for a specific property or user.",
                "operationId": "a15f7d74fd8a01fc0cba02cc1a871267",
                "parameters": [
                    {
                        "name": "reviewable_id",
                        "in": "query",
                        "description": "ID of the item to retrieve reviews for",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "reviewable_type",
                        "in": "query",
                        "description": "Type of item to retrieve reviews for ('property' or 'user')",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "property",
                                "user"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of reviews retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Review"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Submit a review",
                "description": "Submits a review for a property or a user (e.g., a driver).",
                "operationId": "8a6a12f1ab42f8041b6dd9152e7cc7c0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "reviewable_id",
                                    "reviewable_type",
                                    "rating"
                                ],
                                "properties": {
                                    "reviewable_id": {
                                        "description": "ID of the item to review",
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "reviewable_type": {
                                        "description": "Type of item to review ('property' or 'user')",
                                        "type": "string",
                                        "enum": [
                                            "property",
                                            "user"
                                        ],
                                        "example": "property"
                                    },
                                    "rating": {
                                        "type": "integer",
                                        "maximum": 5,
                                        "minimum": 1,
                                        "example": 5
                                    },
                                    "comment": {
                                        "type": "string",
                                        "example": "Excellent service!",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Review submitted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "review": {
                                            "$ref": "#/components/schemas/Review"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Unprocessable Entity (e.g., target not found)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Review": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true,
                        "example": 1
                    },
                    "reviewer_id": {
                        "type": "integer",
                        "example": 12
                    },
                    "reviewable_id": {
                        "description": "ID of the item being reviewed (e.g., property_id, user_id)",
                        "type": "integer",
                        "example": 3
                    },
                    "reviewable_type": {
                        "description": "Model name of the item being reviewed",
                        "type": "string",
                        "example": "App\\Models\\Property"
                    },
                    "rating": {
                        "type": "integer",
                        "maximum": 5,
                        "minimum": 1,
                        "example": 5
                    },
                    "comment": {
                        "type": "string",
                        "example": "It was a wonderful experience!",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}